Simplifying the code in the App.tsx file. Create a hook: useRefreshToken in a separate file and implement it in the application.
