+++ src/components/HomeAdmin/HomeAdmin.tsx
@@ -4,9 +4,11 @@ import {AddHR} from '../AddHR/AddHR';
 import {Header} from '../Header/Header';
 import {useSelector} from "react-redux";
 import {StoreState} from "../../redux-toolkit/store";
+import {useRefreshToken} from "../../utils/useRefreshToken";
 
 export const HomeAdmin = () => {
     const { id } = useSelector((store: StoreState) => store.user);
+    useRefreshToken();
     return (
         <>
             <Header role="admin" id={id}/>
